WIP cont. Enigma class crack method
